# backend/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ANAM-BF - API de la plateforme agrométéo",
    "description": "\nAPI de la plateforme agrométéorologique de l'ANAM-BF.\n\n**Conventions**\n- Format JSON, UTF-8. Dates en UTC au format ISO 8601.\n- Authentification : en-tête `Authorization: Bearer <access_token>` (voir la section *Authentification*).\n  Les contenus publiés (bulletins, alertes, avis, carte, zones) sont **publics** : consultables sans compte.\n- Erreurs : `{\"detail\": \"message lisible\"}` avec le code HTTP adapté (401 non connecté, 403 droits insuffisants,\n  404 introuvable, 409 conflit d'état, 422 données invalides, 429 trop de tentatives).\n- Listes paginées : paramètres `limit` et `offset`, réponse `{\"total\": N, \"items\": [...]}`.\n- Langues : `fr` (français), `en` (anglais), `mos` (mooré).\n- Génération audio/vidéo : asynchrone. Lancer, puis relire le contenu jusqu'à `media.status` = `ready`.\n",
    "version": "1.0.0"
  },
  "paths": {
...
    "/bulletins": {
      "get": {
        "tags": [
          "Bulletins"
        ],
        "summary": "List Bulletins",
        "description": "Bulletins, du plus récent au plus ancien. Sans droits de gestion : uniquement les bulletins publiés.\n\n**Accès :** Public. Une connexion avec droits de gestion donne accès aux contenus non publiés.",
        "operationId": "list_bulletins_bulletins_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "published",
                "draft",
                "archived",
                "all"
              ],
              "type": "string",
              "default": "published",
              "title": "Status"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulletinList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Bulletins"
        ],
        "summary": "Import Bulletin",
        "description": "Importe un bulletin PDF : extraction immédiate du texte, puis génération en arrière-plan des 3 audios\net 3 vidéos (1 à 3 minutes : relire le bulletin jusqu'à `media.status` = `ready`). Le bulletin reste en\nbrouillon jusqu'à `POST /bulletins/{id}/publish`.\n\n**Accès :** Connexion requise, permission `content:manage` (rôles : Agent ANAM, Administrateur).",
        "operationId": "import_bulletin_bulletins_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_import_bulletin_bulletins_post"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulletinOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/bulletins/{bulletin_id}": {
      "get": {
        "tags": [
          "Bulletins"
        ],
        "summary": "Get Bulletin",
        "description": "Détail d'un bulletin : sections du texte, texte lu par langue, PDF, liens des audios et vidéos.\n\n**Accès :** Public. Une connexion avec droits de gestion donne accès aux contenus non publiés.",
        "operationId": "get_bulletin_bulletins__bulletin_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "bulletin_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Bulletin Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulletinOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Bulletins"
        ],
        "summary": "Update Bulletin",
        "description": "**Accès :** Connexion requise, permission `content:manage` (rôles : Agent ANAM, Administrateur).",
        "operationId": "update_bulletin_bulletins__bulletin_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "bulletin_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Bulletin Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulletinUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulletinOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Bulletins"
        ],
        "summary": "Delete Bulletin",
        "description": "Supprime un bulletin non publié (et ses fichiers).\n\n**Accès :** Connexion requise, permission `content:manage` (rôles : Agent ANAM, Administrateur).",
        "operationId": "delete_bulletin_bulletins__bulletin_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "bulletin_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Bulletin Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/bulletins/{bulletin_id}/media": {
      "post": {
        "tags": [
          "Bulletins"
        ],
        "summary": "Regenerate Media",
        "description": "Relance la génération des audios et vidéos (après un échec, ou si les services de traduction ont évolué).\n\n**Accès :** Connexion requise, permission `content:manage` (rôles : Agent ANAM, Administrateur).",
        "operationId": "regenerate_media_bulletins__bulletin_id__media_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "bulletin_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Bulletin Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulletinOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/bulletins/{bulletin_id}/publish": {
      "post": {
        "tags": [
          "Bulletins"
        ],
        "summary": "Publish Bulletin",
        "description": "Publie le bulletin (visible dans l'application) et, selon le paramètre, le diffuse.\n\n**Accès :** Connexion requise, permission `content:publish` (rôles : Agent ANAM, Administrateur).",
        "operationId": "publish_bulletin_bulletins__bulletin_id__publish_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "bulletin_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Bulletin Id"
            }
          },
          {
            "name": "broadcast",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Diffuser tout de suite ; par défaut : paramètre `diffusion.auto_broadcast`",
              "title": "Broadcast"
            },
            "description": "Diffuser tout de suite ; par défaut : paramètre `diffusion.auto_broadcast`"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulletinOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/bulletins/{bulletin_id}/unpublish": {
      "post": {
        "tags": [
          "Bulletins"
        ],
        "summary": "Unpublish Bulletin",
        "description": "Retire le bulletin de l'application (archivé).\n\n**Accès :** Connexion requise, permission `content:publish` (rôles : Agent ANAM, Administrateur).",
        "operationId": "unpublish_bulletin_bulletins__bulletin_id__unpublish_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "bulletin_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Bulletin Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulletinOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/bulletins/{bulletin_id}/broadcast": {
      "post": {
        "tags": [
          "Bulletins"
        ],
        "summary": "Broadcast Bulletin",
        "description": "(Re)diffuse manuellement un bulletin publié. Suivi : `GET /backoffice/broadcasts/{id}`.\n\n**Accès :** Connexion requise, permission `content:publish` (rôles : Agent ANAM, Administrateur).",
        "operationId": "broadcast_bulletin_bulletins__bulletin_id__broadcast_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "bulletin_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Bulletin Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BroadcastOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/bulletins/{bulletin_id}/share-kit": {
      "get": {
        "tags": [
          "Bulletins"
        ],
        "summary": "Share Kit",
        "description": "Texte formaté pour WhatsApp et liens des médias : bouton de diffusion manuelle du back-office.\n\n**Accès :** Connexion requise, permission `content:manage` (rôles : Agent ANAM, Administrateur).",
        "operationId": "share_kit_bulletins__bulletin_id__share_kit_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "bulletin_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Bulletin Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShareKit"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/backoffice/broadcasts/{broadcast_id}": {
      "get": {
        "tags": [
          "Back-office"
        ],
        "summary": "Get Broadcast",
        "description": "Suivi d'une diffusion : décompte par canal et statut, et détail des envois (filtrables). Statut `simulated`\n= envoi simulé (mode développement, rien n'est parti).\n\n**Accès :** Connexion requise, permission `stats:read` (rôles : Responsable communal, Agent ANAM, Administrateur).",
        "operationId": "get_broadcast_backoffice_broadcasts__broadcast_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "broadcast_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Broadcast Id"
            }
          },
          {
            "name": "channel",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Channel"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 200,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BroadcastDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_import_bulletin_bulletins_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "description": "Bulletin PDF de l'ANAM"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_import_bulletin_bulletins_post"
      },
...
      "BroadcastOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "content_type": {
            "type": "string",
            "title": "Content Type"
          },
          "content_id": {
            "type": "string",
            "title": "Content Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "trigger": {
            "type": "string",
            "title": "Trigger"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "target_count": {
            "type": "integer",
            "title": "Target Count"
          },
          "sent_count": {
            "type": "integer",
            "title": "Sent Count"
          },
          "failed_count": {
            "type": "integer",
            "title": "Failed Count"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "type": "object",
        "required": [
          "id",
          "content_type",
          "content_id",
          "status",
          "trigger",
          "created_at",
          "finished_at",
          "target_count",
          "sent_count",
          "failed_count",
          "note"
        ],
        "title": "BroadcastOut"
      },
      "BulletinList": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/BulletinOut"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "total",
          "items"
        ],
        "title": "BulletinList"
      },
      "BulletinOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "date_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date Text"
          },
          "issued_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Issued At"
          },
          "status": {
            "$ref": "#/components/schemas/ContentStatus"
          },
          "parsed": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parsed"
          },
          "texts": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "string"
                },
                "propertyNames": {
                  "enum": [
                    "fr",
                    "en",
                    "mos"
                  ]
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Texts"
          },
          "pdf_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pdf Url"
          },
          "media": {
            "$ref": "#/components/schemas/MediaInfo"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "published_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published At"
          },
          "source": {
            "type": "string",
            "title": "Source"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "date_text",
          "issued_at",
          "status",
          "parsed",
          "texts",
          "pdf_url",
          "media",
          "created_at",
          "published_at",
          "source"
        ],
        "title": "BulletinOut"
      },
      "BulletinUpdate": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "type": "object",
        "title": "BulletinUpdate"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ShareKit": {
        "properties": {
          "text": {
            "additionalProperties": {
              "type": "string"
            },
            "propertyNames": {
              "enum": [
                "fr",
                "en",
                "mos"
              ]
            },
            "type": "object",
            "title": "Text"
          },
          "media": {
            "additionalProperties": {
              "$ref": "#/components/schemas/MediaFiles"
            },
            "propertyNames": {
              "enum": [
                "fr",
                "en",
                "mos"
              ]
            },
            "type": "object",
            "title": "Media"
          },
          "image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Url"
          },
          "note": {
            "type": "string",
            "title": "Note"
          }
        },
        "type": "object",
        "required": [
          "text",
          "media",
          "note"
        ],
        "title": "ShareKit",
        "description": "Contenu prêt à publier manuellement (WhatsApp, réseaux sociaux) : texte formaté et liens médias."
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}